fix(complexities): change float32 to float64 for all result tensors
This avoids a potential error when writing results to a parquet file.
Specifying the complexity metrics in an order in which a float16 tensor
is appended to a float32 tensor in a dataframe and then written to
parquet results in an error as downcasting types seems not possible here.
To be shure to have optimal precision float64 is chosen for all metrics.

+++ aeroblade/src/aeroblade/complexities.py
@@ -68,7 +68,7 @@ def _compute_jpeg(
             nbytes = len(encode_jpeg(convert_image_dtype(patch, torch.uint8), quality=quality))
             patch_results.append(nbytes)
 
-        image_results.append(torch.tensor(patch_results, dtype=torch.float16))
+        image_results.append(torch.tensor(patch_results, dtype=torch.float64))
 
     return torch.stack(image_results) / (patch.shape[1] * patch.shape[2])  # normalize
 
@@ -140,7 +140,7 @@ def _compute_variance(
             patch_variance = np.mean(variance_map)  # Aggregate variance for the patch
             patch_results.append(patch_variance)
 
-        image_results.append(torch.tensor(patch_results, dtype=torch.float32))
+        image_results.append(torch.tensor(patch_results, dtype=torch.float64))
 
     return torch.stack(image_results) / (patch.shape[1] * patch.shape[2])  # normalize
 
@@ -211,7 +211,7 @@ def _compute_meaningful(
 
             patch_results.append(np.sum(complexity))
 
-        image_results.append(torch.tensor(patch_results, dtype=torch.float16))
+        image_results.append(torch.tensor(patch_results, dtype=torch.float64))
 
     return torch.stack(image_results) / (patch.shape[1] * patch.shape[2])  # normalize
